+++ iam.tf
@@ -4,8 +4,20 @@ resource "google_service_account" "nomad" {
 }
 
 # https://registry.terraform.io/providers/hashicorp/google/latest/docs/resources/google_project_iam#google_project_iam_member
-resource "google_project_iam_member" "nomad" {
+resource "google_project_iam_member" "sa_compute_viewer" {
   project = var.project_id
   role    = "roles/compute.viewer"
+  member  = "serviceAccount:${google_service_account.nomad.email}"
+}
+
+resource "google_project_iam_member" "sa_log_writer" {
+  project = var.project_id
+  role    = "roles/logging.logWriter"
+  member  = "serviceAccount:${google_service_account.nomad.email}"
+}
+
+resource "google_project_iam_member" "sa_metric_writer" {
+  project = var.project_id
+  role    = "roles/monitoring.metricWriter"
   member  = "serviceAccount:${google_service_account.nomad.email}"
 }